Added ORM Model for Agency, Implemented basic CRUD functions, Created token-based dependency for authorization
