--- services/auth/app/api/dependencies.py
@@ -1,39 +1,0 @@
-from fastapi import Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer
-from app.services.security import decode_token
-from sqlalchemy.ext.asyncio import AsyncSession
-from app.db.deps import get_db
-from app.models.user import RoleEnum, User
-from app.crud.crud_user import get_user_by_email
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")
-
-async def get_current_user(
-    token: str = Depends(oauth2_scheme),
-    db: AsyncSession = Depends(get_db),
-) -> User:
-    try:
-        payload = decode_token(token)
-    except Exception:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid token",
-        )
-    
-    user = await get_user_by_email(db, payload["email"])
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail=f"User not found: {payload['email']}",
-        )
-    return user
-
-def require_role(role: RoleEnum):
-    def role_checker(user: User = Depends(get_current_user)):
-        if user.role != role:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail="Not enough permissions"
-            )
-        return user
-    return role_checker
