--- services/auth/app/crud/crud_user.py
@@ -1,19 +1,0 @@
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.future import select
-from app.models.user import User
-from app.schemas.user import UserCreate
-from app.services.security import hash_password
-
-async def get_user_by_email(db: AsyncSession, email: str) -> User | None:
-    result = await db.execute(select(User).where(User.email == email))
-    return result.scalars().first()
-
-async def create_user(db: AsyncSession, user_in: UserCreate) -> User:
-    user = User(
-        email=user_in.email,
-        hashed_password=hash_password(user_in.password)
-    )
-    db.add(user)
-    await db.commit()
-    await db.refresh(user)
-    return user
