--- services/reconciliation/pyproject.toml
@@ -1,33 +1,0 @@
-[project]
-name = "innotour-backend-Recon"
-version = "0.1.0"
-description = "Microservices: Reconciliation Service"
-authors = [
-    {name = "Daniil Vasilev",email = "creepydanunity@gmail.com"}
-]
-license = {text = "MIT"}
-readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-    "fastapi (>=0.115.12,<0.116.0)",
-    "uvicorn[standard] (>=0.34.2,<0.35.0)",
-    "sqlalchemy (>=2.0.41,<3.0.0)",
-    "alembic (>=1.16.0,<2.0.0)",
-    "pydantic (>=2.11.4,<3.0.0)",
-    "python-dotenv (>=1.1.0,<2.0.0)"
-]
-
-
-[build-system]
-requires = ["poetry-core>=2.0.0,<3.0.0"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.3.5"
-pytest-asyncio = "^0.26.0"
-black = "^25.1.0"
-isort = "^6.0.1"
-flake8 = "^7.2.0"
-mypy = "^1.15.0"
-bandit = "^1.8.3"
-
