+++ services/scheduling/app/schemas/error.py
