+++ services/user/alembic/env.py
@@ -1,0 +1,91 @@
+from logging.config import fileConfig
+import os, sys, asyncio
+
+from sqlalchemy import pool
+from sqlalchemy.ext.asyncio import create_async_engine
+from alembic import context
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+
+from app.core.config import settings
+from app.db.base import Base
+import app.models.user
+import app.models.token
+
+# this is the Alembic Config object, which provides
+# access to the values within the .ini file in use.
+config = context.config
+config.set_main_option("sqlalchemy.url", str(settings.database_url))
+# Interpret the config file for Python logging.
+# This line sets up loggers basically.
+if config.config_file_name is not None:
+    fileConfig(config.config_file_name)
+
+# add your model's MetaData object here
+# for 'autogenerate' support
+# from myapp import mymodel
+# target_metadata = mymodel.Base.metadata
+target_metadata = Base.metadata
+
+# other values from the config, defined by the needs of env.py,
+# can be acquired:
+# my_important_option = config.get_main_option("my_important_option")
+# ... etc.
+
+
+def run_migrations_offline() -> None:
+    """Run migrations in 'offline' mode.
+
+    This configures the context with just a URL
+    and not an Engine, though an Engine is acceptable
+    here as well.  By skipping the Engine creation
+    we don't even need a DBAPI to be available.
+
+    Calls to context.execute() here emit the given string to the
+    script output.
+
+    """
+    url = config.get_main_option("sqlalchemy.url")
+    assert url is not None, "Missing sqlalchemy.url"
+    context.configure(
+        url=url,
+        target_metadata=target_metadata,
+        literal_binds=True,
+        dialect_opts={"paramstyle": "named"},
+    )
+
+    with context.begin_transaction():
+        context.run_migrations()
+
+
+def run_migrations_online() -> None:
+    """Run migrations in 'online' mode.
+
+    In this scenario we need to create an Engine
+    and associate a connection with the context.
+
+    """
+    url = config.get_main_option("sqlalchemy.url")
+    assert url is not None, "Missing sqlalchemy.url"
+
+    connectable = create_async_engine(
+        url,
+        poolclass=pool.NullPool,
+    )
+
+    def do_run_migrations(sync_conn):
+        context.configure(connection=sync_conn, target_metadata=target_metadata)
+        with context.begin_transaction():
+            context.run_migrations()
+
+    async def _runner():
+        async with connectable.connect() as conn:
+            await conn.run_sync(do_run_migrations)
+
+    asyncio.run(_runner())
+
+
+if context.is_offline_mode():
+    run_migrations_offline()
+else:
+    run_migrations_online()
