+++ services/user/app/core/exceptions.py
@@ -1,0 +1,60 @@
+from fastapi import HTTPException, status
+
+
+class EmailAlreadyRegistered(HTTPException):
+    def __init__(self, email: str):
+        super().__init__(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"Email {email} is already registered"
+        )
+        
+        self.email = email
+        
+class InvalidCredentials(HTTPException):
+    def __init__(self):
+        super().__init__(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password"
+        )
+
+class TokenExpired(HTTPException):
+    def __init__(self):
+        super().__init__(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Refresh token scope is incorrect"
+        )
+
+class TokenInvalid(HTTPException):
+    def __init__(self):
+        super().__init__(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Refresh token is invalid"
+        )
+
+class TokenWrongScope(HTTPException):
+    def __init__(self):
+        super().__init__(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Refresh token scope is invalid"
+        )
+
+class CSRFInvalid(HTTPException):
+    def __init__(self):
+        super().__init__(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="CSRF validation failed"
+        )
+
+class UserNotFound(HTTPException):
+    def __init__(self):
+        super().__init__(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="User not found"
+        )
+
+class PermissionRequired(HTTPException):
+    def __init__(self):
+        super().__init__(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not enough permissions"
+        )
