+++ services/user/app/db/session.py
@@ -1,0 +1,8 @@
+from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession, async_sessionmaker
+
+from app.core.config import settings
+
+
+DATABASE_URL = str(settings.database_url)
+engine = create_async_engine(DATABASE_URL, future=True, echo=False)
+AsyncSessionLocal = async_sessionmaker(bind=engine, expire_on_commit=False, class_=AsyncSession)
