+++ services/user/app/models/token.py
@@ -1,0 +1,41 @@
+import hashlib
+import uuid
+from datetime import datetime
+
+from sqlalchemy import DateTime, String, ForeignKey, func
+from sqlalchemy.orm import Mapped, mapped_column
+
+from app.db.base import Base
+
+class RefreshToken(Base):
+    __tablename__ = "refresh_tokens"
+
+    id: Mapped[uuid.UUID] = mapped_column(
+        primary_key=True,
+        default=uuid.uuid4,
+        index=True,
+    )
+    user_id: Mapped[int] = mapped_column(
+        ForeignKey("users.id", ondelete="CASCADE"),
+        nullable=False,
+        index=True,
+    )
+    token_hash: Mapped[str] = mapped_column(
+        String,
+        unique=True,
+        index=True,
+        nullable=False,
+    )
+    created_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        server_default=func.now(),
+        nullable=False,
+    )
+    expires_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        nullable=False,
+    )
+
+    @classmethod
+    def hash_token(cls, raw_token: str) -> str:
+        return hashlib.sha256(raw_token.encode()).hexdigest()
