+++ services/user/app/schemas/error.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class ErrorResponse(BaseModel):
+    detail: str
